Patch 1
I realised through the help of one of my friends that there's an issue with the encryption and decryption parser! It was a stupid mistake on my behalf as I didn't properly do a full conversion of the code from Harbinger v1, I apologise for the disfunction of the program.